Theia specific temporary fix: floor character width when computing number of columns
Try to fix this issue: https://github.com/theia-ide/theia/issues/735

xterm is using the non-floored character width (e.g. 7.8125px) when
computing the number of columns that fit in the available space.
However, it computes its canvas size using that same number of columns,
but the floored character width (7px).  This results in less columns
than would fit.

As a temporary fix, made proposeGeometry use the floored character width
when computing the number of columns.  It works for me, but I don't
think it would play well with HiDPI displays, since the real character
width is scaled by the devicePixelRatio before flooring.  I don't have
a device with HiDPI to test.